Fix for RoundCube (/script/fixwebmail). Script fixwebmail never ever worked in all LxAdmin/Kloxo versions. Related to issue #421

+++ kloxo/bin/misc/lxinstall_hordegroupware_db.php
@@ -4,11 +4,14 @@
 // Beta Version 29-jan-2011
 //
 //
+// - It always resets the webmail-database/webmail-user passwords when running this script.
+//   Just a little bit security :) 
 //
+
 
 print("Start fixing webmail\n\n");
 
-print("Fixing Horde......\n###\n");
+print("###\nFixing Horde...\n");
 
 include_once "htmllib/lib/include.php"; 
 
@@ -31,32 +34,35 @@ $result = mysql_query("CREATE DATABASE `horde_groupware`");
 
 if (!$result)
 {
+ $skip = false;
  print("Database already exists. Check for tables\n");
  $result = mysql_select_db('horde_groupware');
  if (!$result) { print("Something went wrong, can not select horde database!\n"); }
 
  $query = "SELECT `user_uid` FROM `horde_users` WHERE `user_uid`='admin'"; 
  $result = mysql_query($query);
- if (!$result) {
-  print("Something went wrong, could not find admin user\n");
- } else {
+ if ($result) {
   print("Your Database looks fine. No need to fix\n");
-  exit;
- }
+  $skip = true;
+  } else {
+  $skip = false;
+  print("Something went wrong, could not find admin user\n");
+ 
 
- print("Dropping database\n");
- $result = mysql_query("DROP DATABASE `horde_groupware`");
- if (!$result)
- {
-  print("Could not drop horde database!\nScript Abort\n\n");
-  exit;
- }
+  print("Dropping database\n");
+  $result = mysql_query("DROP DATABASE `horde_groupware`");
+  if (!$result)
+  {
+   print("Could not drop horde database!\nScript Abort\n\n");
+   exit;
+  }
 
- print("Create database\n");
- $result = mysql_query("CREATE DATABASE `horde_groupware`");
- if (!$result) {
-  print("There is REALY something very very wrong... Go to http://forum.lxcenter.org/ and report.\n\n");
-  exit;
+  print("Create database\n");
+  $result = mysql_query("CREATE DATABASE `horde_groupware`");
+  if (!$result) {
+   print("There is REALY something very very wrong... Go to http://forum.lxcenter.org/ and report.\n\n");
+   exit;
+  }
  }
 
 }
@@ -76,6 +82,9 @@ print("Granting privileges\n");
 $result = mysql_query("GRANT ALL ON horde_groupware.* TO horde_groupware@localhost IDENTIFIED BY '$pass'", $link);
 mysql_query("flush privileges", $link);
 if (!$result) { print("Could not grant privileges\nScript Abort.\n"); exit; }
+print("Database connection is fixed\n");
+
+if (!$skip) {
 	
 print("Fix database values in horde sql importfile\n");
 
@@ -108,11 +117,32 @@ lfile_put_contents($hordefile, $content);
 	$query = "INSERT INTO horde_users (user_uid, user_pass) VALUES ('admin','21232f297a57a5a743894a0e4a801fc3')" ;
 	$result = mysql_query($query);
         if (!$result) { print("Something went wrong, could not add admin user into horde database\n"); }
+}
 
+print("###\nFixing RoundCube...\n");
+print("Just fixing RoundCube's database connection!!\n");
 
-print("###\nFixing RoundCube\n");
-print("Sorry, roundcubefix is not created yet....\n\n");
+$result = mysql_select_db('roundcubemail');
+if (!$result) { print("Something went wrong, can not select RoundCube database!\n"); exit; }
 
+print("Generating password..\n");
+$pass = randomString(8);
+dprint("Generated Pass ".$pass."\n");
 
-print("Done\n");
+print("Add Password to configfile\n");
+$roundcubefileIN = "/usr/local/lxlabs/kloxo/file/webmail-chooser/db.inc.phps";
+$roundcubefileOUT = "/home/kloxo/httpd/webmail/roundcube/config/db.inc.php";
+$content = file_get_contents($roundcubefileIN);
+$content = str_replace("mysql://roundcube:pass", "mysql://roundcube:" . $pass, $content);
+print("Remove system readonly attribute from configfile\n");
+system("chattr -i /home/kloxo/httpd/webmail/roundcube/config/db.inc.php");
+lfile_put_contents($roundcubefileOUT, $content);
+
+print("Granting privileges\n");
+$result = mysql_query("GRANT ALL ON roundcubemail.* TO roundcube@localhost IDENTIFIED BY '$pass'", $link);
+mysql_query("flush privileges", $link);
+if (!$result) { print("Could not grant privileges\nScript Abort.\n"); exit; }
+print("Database connection is fixed\n");
+
+print("\n###\nDone\n");
 exit;
